LazySupplier as a subclass of F
